python-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
google-generativeai==0.3.1
marshmallow==3.20.1
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
responses==0.24.1
gunicorn==21.2.0
